Bug 572623 redesign license containment
Merge 'floating' feature to 'licensing': redirect 'lic' components
