+++ bundles/org.eclipse.passage.loc.floating.emfforms/META-INF/MANIFEST.MF
@@ -32,7 +32,4 @@ Require-Bundle: javax.inject;bundle-version="0.0.0",
  org.eclipse.passage.lic.emf;bundle-version="0.0.0",
  org.eclipse.passage.lic.floating.model;bundle-version="0.1.0",
  org.eclipse.passage.loc.workbench.emfforms;bundle-version="0.0.0"
-Service-Component: OSGI-INF/org.eclipse.passage.loc.floating.emfforms.EnvironmentRendererService.xml,
- OSGI-INF/org.eclipse.passage.loc.floating.emfforms.AuthenticationExpressionRendererService.xml,
- OSGI-INF/org.eclipse.passage.loc.floating.emfforms.VersionMatchRendererService.xml
 Bundle-ActivationPolicy: lazy
