+++ bundles/org.eclipse.passage.loc.licenses.core/META-INF/MANIFEST.MF
@@ -17,8 +17,8 @@ Require-Bundle: org.eclipse.osgi.services;bundle-version="0.0.0",
  org.eclipse.passage.loc.api;bundle-version="0.0.0",
  org.eclipse.passage.loc.products.core;bundle-version="0.6.0",
  org.eclipse.passage.loc.users.core;bundle-version="1.0.0",
- org.eclipse.passage.lic.floating.model;bundle-version="0.1.0",
- org.eclipse.passage.lic.base;bundle-version="1.0.100"
+ org.eclipse.passage.lic.base;bundle-version="1.0.100",
+ org.eclipse.passage.lic.floating;bundle-version="0.1.1"
 Export-Package: org.eclipse.passage.loc.internal.licenses;
   x-friends:="org.eclipse.passage.loc.licenses.emfforms,
    org.eclipse.passage.loc.licenses.ui,
