+++ bundles/org.eclipse.passage.loc.licenses.emfforms/OSGI-INF/org.eclipse.passage.loc.licenses.emfforms.renderers.VersionMatchRendererService.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.passage.loc.licenses.emfforms.renderers.VersionMatchRendererService">
+   <service>
+      <provide interface="org.eclipse.emfforms.spi.swt.core.di.EMFFormsDIRendererService"/>
+   </service>
+   <reference bind="bindEMFFormsDatabinding" interface="org.eclipse.emfforms.spi.core.services.databinding.EMFFormsDatabinding" name="EMFFormsDatabinding" unbind="unbindEMFFormsDatabinding"/>
+   <reference bind="bindReportService" interface="org.eclipse.emfforms.spi.common.report.ReportService" name="ReportService" unbind="unbindReportService"/>
+   <implementation class="org.eclipse.passage.loc.licenses.emfforms.renderers.VersionMatchRendererService"/>
+</scr:component>
